refactor(api): move find certification centers method to subdomain repository
