+++ api/lib/domain/usecases/find-paginated-filtered-certification-centers.js
@@ -1,5 +1,9 @@
-const findPaginatedFilteredCertificationCenters = function ({ filter, page, certificationCenterRepository }) {
-  return certificationCenterRepository.findPaginatedFiltered({ filter, page });
+const findPaginatedFilteredCertificationCenters = function ({
+  filter,
+  page,
+  organizationalEntitiesCertificationCenterRepository,
+}) {
+  return organizationalEntitiesCertificationCenterRepository.findPaginatedFiltered({ filter, page });
 };
 
 export { findPaginatedFilteredCertificationCenters };
